Move ThreadEventQueue to its own file.
